+++ src/index.ts
@@ -15,5 +15,6 @@ export { Scenario } from './Scenario';
 export { Messenger } from './Messenger/Adapter';
 export { Telegram } from './Telegram/Adapter';
 export { Response as IResponse } from './Response';
+export { MessageMiddleware } from './Middlewares/Message';
 export { IEvent } from './Events/Event';
 export * from './Events/SceneEvents';